refactor: shared folder
