+++ apps/angular-feature-sliced/src/pages/task-details/task-details.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
-import { SharedModule, UiKitModule } from 'shared';
 import { ToggleTaskModule } from 'features/toggle-task';
 import { TaskModule } from 'entities/task';
+import { SharedModule } from 'shared/lib';
+import { UiKitModule } from 'shared/ui';
 
 import { TaskDetailsRoutingModule } from './task-details-routing.module';
 import { TaskDetailsPage } from './task-details.page';
