+++ apps/angular-feature-sliced/src/pages/tasks-list/task-list.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
-import { SharedModule, UiKitModule } from 'shared';
 import { TasksFilterModule } from 'features/tasks-filter';
 import { ToggleTaskModule } from 'features/toggle-task';
 import { TaskModule } from 'entities/task';
+import { SharedModule } from 'shared/lib';
+import { UiKitModule } from 'shared/ui';
 
 import { TasksListPage } from './tasks-list.page';
