--- apps/angular-feature-sliced/src/shared/index.ts
@@ -1,2 +1,0 @@
-export { UiKitModule } from './ui-kit.module';
-export { SharedModule } from './shared.module';
