+++ apps/angular-feature-sliced/src/shared/lib/index.ts
@@ -1,2 +1,3 @@
 export * as RouterModel from './router';
 export { RouterFacade } from './router';
+export { SharedModule } from './shared.module';
